requirements.txt
python-dotenv==1.0.0
google-generativeai>=0.3.0
pymongo>=4.6.0
requests>=2.31.0